requirements-docker.txt
bill==0.0.1
blinker==1.7.0
catboost==1.2.3
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
Flask==3.0.2
fonttools==4.50.0
graphviz==0.20.2
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
## !! Could not determine repository location
# -e c:\users\jogi\projects\ml-project-main
MarkupSafe==2.1.5
matplotlib==3.8.3
mysql-connector-python==8.3.0
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pillow==10.2.0
plotly==5.20.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
tenacity==8.2.3
threadpoolctl==3.3.0
tzdata==2024.1
Werkzeug==3.0.1
xgboost==2.0.3